requirements.txt
# Core: pip install -r requirements.txt then run: .venv/bin/uvicorn main:app --reload --port 8000
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0

# Database
supabase==2.9.0

# Auth & Security
python-jose[cryptography]==3.3.0
python-multipart==0.0.6

# HTTP
httpx==0.27.0
requests==2.31.0

# AI
openai==1.10.0

# Data (pydantic 2.10+ has Python 3.13 wheels)
pydantic>=2.10.0,<3

# Utilities
python-dateutil==2.8.2
Pillow>=11.0.0
